backend/requirements.txt
fastapi==0.116.1
google-adk==1.19.0
google-genai==1.25.0
pydantic==2.10.1
pydantic-settings==2.10.1
python-dotenv==1.0.1
sentry-sdk==2.33.0
uvicorn==0.35.0